+++ app/package.json
@@ -20,8 +20,6 @@
     "eslint": "8.46.0",
     "eslint-config-next": "13.4.12",
     "ethers": "5.7.0",
-    "highlight.js": "^11.8.0",
-    "highlightjs-solidity": "^2.0.6",
     "next": "13.4.12",
     "postcss": "8.4.27",
     "react": "18.2.0",
